logaut and useWindowSize
